% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows_fill.R
\name{bind_rows_fill}
\alias{bind_rows_fill}
\title{Bind Rows and Fill}
\usage{
bind_rows_fill(lst, id_name = "file")
}
\arguments{
\item{lst}{A list of data frames to bind.}

\item{id_name}{Name of the identifier column to add to the output tibble.}
}
\value{
A tibble with missing columns filled and an .id column added.
}
\description{
Bind rows while filling missing columns; adds .id column name supplied
}
\keyword{internal}
