% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_multistate_histories.R
\name{build_multistate_histories}
\alias{build_multistate_histories}
\title{Build capture histories for a multi-state model using \code{marked}}
\usage{
build_multistate_histories(ch_data)
}
\arguments{
\item{ch_data}{Data frame or tibble with at least the following columns:
\describe{
\item{\code{tag_code}}{Unique identifier for each tagged individual.}
\item{\code{ch}}{Encounter history string (e.g., "11001"). The value
1 = released/detected, 0 = not observed, and a 2 = detected but removed and not available for observation at the next site.}
}}
}
\description{
Converts capture histories from \code{build_capture_histories} to a format
usable for multi-state modeling in package \code{marked}.
}
\author{
Ryan N. Kinzer
}
