% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DART_site_config.R
\name{DART_site_config}
\alias{DART_site_config}
\title{Load PTAGIS interrogation site configuration}
\usage{
DART_site_config(configDate = Sys.Date())
}
\arguments{
\item{configDate}{last day of configuration}
}
\value{
A tibble of site observations.
}
\description{
Download site configuration file from DART (https://www.cbr.washington.edu/paramest/docs/pitpro/updates/sites_config.txt) and format as a data frame.
}
\examples{
\dontrun{
# recommended: store your key once per session
Sys.setenv(PTAGIS_API_KEY = "YOUR-KEY-HERE")
get_tag_history(tag_code = "384.1B79726A98")

# or pass the key manually, and reduce fields
get_tag_history(api_key = "YOUR-KEY-HERE", tag_code = "384.1B79726A98",
                fields = c("tag_code", "site_code", "event_type", "event_date"))
}

}
\author{
Ryan Vosbigian
}
