% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mrr_files.R
\name{download_mrr_files}
\alias{download_mrr_files}
\title{Download & Parse Multiple MRR .xml Files}
\usage{
download_mrr_files(filenames)
}
\arguments{
\item{filenames}{}
}
\value{
a named list; each element is the parse list from \verb{get_file_data(..., return = "list}
}
\description{
Download & Parse Multiple MRR .xml Files
}
\author{
Ryan Kinzer
}
