% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_marked_cjs_multiple.R
\name{fit_marked_cjs_multiple}
\alias{fit_marked_cjs_multiple}
\title{Fit a CJS Model Using \code{marked}}
\usage{
fit_marked_cjs_multiple(
  ch_data,
  site_mapping,
  phi_formula = ~time * release_group,
  p_formula = ~time * detect_site,
  hessian = TRUE,
  conf_level = 0.95,
  ...
)
}
\arguments{
\item{ch_data}{Data frame with \code{tag_code}, \code{ch}, \code{release_group} and optional
individual covariates.}

\item{site_mapping}{Data frame with \code{release_group}, \code{detect_site}, and \code{occ}.}

\item{phi_formula}{Formula for survival. Default \code{~ time*release_group}.}

\item{p_formula}{Formula for detection. Default \code{~ time*detect_site}.}

\item{hessian}{Logical; passed to \code{marked::crm()}.}

\item{conf_level}{Confidence level for intervals.}

\item{...}{Additional arguments passed to \code{marked::crm()}.}
}
\value{
List with \code{model}, \code{phi}, \code{cum_phi}, \code{p},
\code{plots}, and \code{covariance_mode}.
}
\description{
Fits a Cormack-Jolly-Seber model for multiple release groups using \code{marked} and returns survival,
cumulative survival, and detection estimates. Detection probability is estimated jointly for all release groups
and survival is estimated independently.
}
\author{
Ryan N. Kinzer

Michelle A. Briggs
}
