% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_marray_cjs_multiple.R
\name{fit_marray_cjs_multiple}
\alias{fit_marray_cjs_multiple}
\title{Fit a CJS Model Using an M-array}
\usage{
fit_marray_cjs_multiple(marray_list)
}
\arguments{
\item{m_array_list}{A tibble with a release_group column, containing the release group or location
and an m_array column, containing data from \code{build_marray()}.}
}
\value{
List with \code{phi}, \code{p}, \code{fit}, and \code{st_err}.
}
\description{
Fits a CJS-style likelihood from a Burnham-style M-array for multiple release groups using \code{RTMB}. The model returns survival
and detection probability estimates. Detection probability is estimated jointly for all release groups
and survival is estimated independently.
}
\author{
Michelle A. Briggs
}
