% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interrogation_sites.R
\name{get_interrogation_sites}
\alias{get_interrogation_sites}
\title{List interrogation sites (metadata)}
\usage{
get_interrogation_sites(active_only = TRUE)
}
\arguments{
\item{active_only}{logical; return only active/currently reporting sites.}
}
\value{
A tibble of site metadata.
}
\description{
List interrogation sites (metadata)
}
