% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mrr_files.R
\name{get_mrr_files}
\alias{get_mrr_files}
\title{Retrieve File Info for Project Code and Year}
\usage{
get_mrr_files(code, year = NULL, page = 1, page_size = 1000, all_pages = TRUE)
}
\arguments{
\item{code}{A single PTAGIS project code (3 characters), e.g. "CDR","LGR", etc.}

\item{year}{A four digit year as an integer.}

\item{page, page_size}{Integers controlling pagination when \code{all_pages = FALSE}.}

\item{all_pages}{Logical; if \code{TRUE} (default), the function automatically
retrieves and combines all pages of results.}
}
\value{
A tibble containing metadata for submitted files.
}
\description{
Retrieve file information submitted by a
mark-recapture-recovery (MRR) project code within a given year.
}
\examples{
\dontrun{ get_mrr_files(code = "LGR", year = 2024) }

}
\author{
Ryan Kinzer
}
