% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_codes.R
\name{get_project_codes}
\alias{get_project_codes}
\title{Download PTAGIS Project Codes}
\usage{
get_project_codes()
}
\value{
A character vector of available project codes (uppercase, sorted, unique).
}
\description{
Download a character vector of project codes available in PTAGIS. Project
codes are used to identify the mark-recapture-recovery (MRR) project under which the
data was submitted. More information on available project codes can be found
[here](https://www.ptagis.org/Resources/ValidationCodes?domainFilter=MrrProject).
}
\examples{
\dontrun{ get_project_codes() }
}
\author{
Ryan Kinzer
}
