% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_years.R
\name{get_project_years}
\alias{get_project_years}
\title{Retrieve Years for a PTAGIS Project}
\usage{
get_project_years(code)
}
\arguments{
\item{code}{A single PTAGIS project code (3 characters), e.g. "LGR".}
}
\value{
An integer vector of available years (sorted, unique).
}
\description{
Retrieves the years for which data are available for a given
PTAGIS project code.
}
\examples{
\dontrun{ get_project_years("LGR") }

}
\author{
Ryan Kinzer
}
