% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_data.R
\name{parse_mrr_xml}
\alias{parse_mrr_xml}
\title{Parse a PTAGIS MRR xml_document into tidy tibbles}
\usage{
parse_mrr_xml(doc)
}
\description{
Parse a PTAGIS MRR xml_document into tidy tibbles
}
\keyword{internal}
