% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot survival by reach (base graphics placeholder)}
\usage{
plot_survival(surv_tbl)
}
\arguments{
\item{surv_tbl}{table returned by \code{summarize_survival()}.}
}
\description{
Plot survival by reach (base graphics placeholder)
}
