% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marked_helpers.R
\name{print.fit_marked_cjs}
\alias{print.fit_marked_cjs}
\title{Quick printer for fit_marked_cjs output}
\usage{
\method{print}{fit_marked_cjs}(x, ...)
}
\arguments{
\item{x}{object returned by fit_marked_cjs}
}
\description{
Quick printer for fit_marked_cjs output
}
