% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fit_marked_cjs.R
\name{print_fit_marked_cjs}
\alias{print_fit_marked_cjs}
\title{Quick printer for fit_marked_cjs output}
\usage{
print_fit_marked_cjs(x, ...)
}
\arguments{
\item{x}{object returned by fit_marked_cjs}

\item{...}{additional arguments (unused)}
}
\description{
Quick printer for fit_marked_cjs output
}
