% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptagis_GET.R
\name{ptagis_GET}
\alias{ptagis_GET}
\title{Internal: HTTP GET with retries for PTAGIS}
\usage{
ptagis_GET(path, query = list(), base_url = "https://api.ptagis.org", ...)
}
\description{
Internal: HTTP GET with retries for PTAGIS
}
\keyword{internal}
