% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_event_groups.R
\name{save_event_groups}
\alias{save_event_groups}
\title{Subset events and output files based on defined groups.}
\usage{
save_event_groups(
  events,
  group_cols,
  event_type_value = c("Mark", "Recapture", "Tally", "Recovery", "Passive Recapture"),
  out_dir = NULL,
  remove_blank_tags = TRUE,
  save_TagIds = TRUE,
  unique_tags = TRUE
)
}
\arguments{
\item{events}{A single data.frame (or Tibble) object.}

\item{group_cols}{Character vector of grouping variables as character strings}

\item{event_type_value}{Character; one of \code{"Mark"} (default), \code{"Recapture"}, \code{"Tally"},\code{"Recovery"},\code{"Passive Recapture"},}

\item{out_dir}{Character string of desired output directory}

\item{remove_blank_tags}{Logical; if \code{TRUE} events containing \code{NULL} PIT-tag codes (i.e., dot-outs) will be removed from the returned \code{.rds} file.}

\item{save_TagIds}{Logical; if \code{TRUE} additional \code{.txt} files will be created
for each defined group containing only PIT-tag codes.}

\item{unique_tags}{Logical; if \code{TRUE} if you want duplicate PIT-tag codes be removed from the \code{.txt} list.}
}
\description{
Splits a large event object into groups based on \code{event_type} and
user defined groups within the data. After splitting each group is saved to
individual \code{.rds} files, and the user has the option to save an
additional \code{.txt} file of PIT-tag codes for easy upload into PTAGIS
queries (e.g., complete tag histories, interogation details).
}
\author{
Ryan N. Kinzer
}
