% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_marray.R
\name{build_marray}
\alias{build_marray}
\title{Create an M-array of releases, detections, and removals.}
\usage{
build_marray(ch_data, locs_def = NULL)
}
\arguments{
\item{ch_data}{Data frame or tibble with at least the following columns:
\describe{
\item{\code{tag_code}}{Unique identifier for each tagged individual.}
\item{\code{ch}}{Encounter history string (e.g., "11001"). The value
1 = released/detected, 0 = not observed, and a 2 = detected but removed and not available for observation at the next site.}
}}

\item{locs_def}{Either a character vector defining the ordered occasions
(e.g., \code{c("SECTRP","ZEN",...)}), or a named list mapping one or more site codes
to occasions
(e.g., \code{list(LGR = c("GRJ","GRS"), Down = c("LMN","MCN","BON"))}).}
}
\value{
a data.frame
}
\description{
Builds a matrix of marks released, observed, and never seen.
}
\author{
Ryan N. Kinzer
}
