% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DART_site_config.R
\name{DART_site_config}
\alias{DART_site_config}
\title{Load interrogation site configuration file.}
\usage{
DART_site_config(configDate = Sys.Date())
}
\arguments{
\item{configDate}{last day of configuration}
}
\value{
A tibble of site observations.
}
\description{
Download site configuration file from DART (https://www.cbr.washington.edu/paramest/docs/pitpro/updates/sites_config.txt) and format as a data frame. The function was originally written by Ryan Vosbigian and contained in the 'https://github.com/ryanvosbigian/space4time' R package.
}
\author{
Ryan N. Kinzer
}
