% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mrr_files.R
\name{download_mrr_files}
\alias{download_mrr_files}
\title{Download and Parse Multiple MRR XML Files}
\usage{
download_mrr_files(filenames, drop_pdvs = FALSE)
}
\arguments{
\item{filenames}{Character vector of MRR file names to download and process.}
}
\value{
A named list where each element corresponds to one input file.
Each element contains the parsed output from \code{get_file_data(..., return = "list")}.
}
\description{
Downloads and parses multiple mark-recapture-recovery (MRR) XML files.
Each file is read using \code{get_file_data()} and returned as a list.
}
\author{
Ryan Kinzer
}
