% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_marray_cjs.R
\name{fit_marray_cjs}
\alias{fit_marray_cjs}
\title{Fit a CJS Model Using an M-array.}
\usage{
fit_marray_cjs(m_array)
}
\arguments{
\item{m_array}{Data frame or tibble produced by \code{build_array}.}
}
\description{
Fits a Cormack-Jolly-Seber (CJS) capture-recapture model using the M-array approach
as described in Burnham et al. 1987. The approach allows for censored data
(i.e., fish removed at a detection site for sampling or transportion). The
function returns tidy results for survival (\code{Phi}), detection (\code{p}),
and lambda (i.e., the product of survival and detection at the last reach and site).

Cumulative survival estimates are computed with covariance-aware confidence intervals
when possible; otherwise, an independence-based fallback is used.
}
\author{
Ryan N. Kinzer
}
