% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_data.R
\name{get_file_data}
\alias{get_file_data}
\title{Download and Parse a MRR File}
\usage{
get_file_data(filename, return = c("list", "xml", "session", "events"))
}
\arguments{
\item{filename}{name of the file to be downloaded as character
e.g. "CDR-2024-072-JCT.xml".}

\item{return}{the format of the object to be returned:
"list" (default), "xml", "session", or "events"}
}
\value{
if "list", a list with tibbles: session, events, session_pdv_fields,
detail_pdv_fields
}
\description{
Download and parse data within a single PTAGIS mark-recapture-recovery
(MRR) .xml file.
}
