% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interrogation_sites.R
\name{get_interrogation_sites}
\alias{get_interrogation_sites}
\title{List Interrogation Sites}
\usage{
get_interrogation_sites(active_only = TRUE)
}
\arguments{
\item{active_only}{Logical; if \code{TRUE} (default), return only active or currently
reporting interrogation sites. If \code{FALSE}, return all interrogation sites.}
}
\value{
A tibble containing interrogation site metadata.
}
\description{
Retrieves metadata for interrogation sites from PTAGIS. The function can return either
all registered sites (past and present) or only those active and contributing data.
}
\author{
Ryan Kinzer
}
