% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_codes.R
\name{get_project_codes}
\alias{get_project_codes}
\title{Retrieve PTAGIS Project Codes}
\usage{
get_project_codes()
}
\value{
A character vector of available project codes (uppercase, sorted, unique).
}
\description{
Retrieves a character vector of all mark-recapture-recovery (MRR)
project codes currently available in the \href{https://www.ptagis.org}{PTAGIS} database.
Project codes identify the MRR project under which PIT-tag data were submitted.

For a full list of available project codes with names, years, and status, see
the \href{https://www.ptagis.org/Resources/ValidationCodes?domainFilter=MrrProject}{PTAGIS Validation Codes page}.
}
\examples{
\dontrun{ get_project_codes() }

}
\author{
Ryan Kinzer
}
