% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_codes.R
\name{get_project_codes}
\alias{get_project_codes}
\title{Download available PTAGIS project codes}
\usage{
get_project_codes()
}
\value{
A character vector of available project codes (uppercase, sorted, unique).
}
\description{
Download available PTAGIS project codes
}
\examples{
\dontrun{ get_project_codes() }
}
