% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_curve.R
\name{passage_curve}
\alias{passage_curve}
\title{Generate cumulative passage curve data}
\usage{
passage_curve(timing)
}
\arguments{
\item{timing}{list returned by \code{fit_timing()}.}
}
\value{
data.frame with date and cumulative proportion.
}
\description{
Generate cumulative passage curve data
}
