% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing_helpers.R
\name{plot_arrival_travel}
\alias{plot_arrival_travel}
\title{Plot arrival timing (ECDF) and travel times by leg}
\usage{
plot_arrival_travel(timing_list, tz = "UTC")
}
\arguments{
\item{timing_list}{Result from `summarize_arrival_travel()`}

\item{tz}{Label/timezone for axis formatting (default "UTC")}
}
\value{
A list with `$arrival_ecdf` and `$travel_time` plot objects.
  If ggplot2 is not available, returns functions that draw base plots when called.
}
\description{
Plot arrival timing (ECDF) and travel times by leg
}
