% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fit_marked_cjs.R
\name{print_fit_marked_cjs}
\alias{print_fit_marked_cjs}
\title{Print CJS Model Summary}
\usage{
print_fit_marked_cjs(z, ...)
}
\arguments{
\item{z}{An object returned by \code{fit_marked_cjs()}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{z}.
}
\description{
Provides a concise display of results from \code{fit_marked_cjs()}, including
the model name, the first several survival (\code{phi}) estimates, and the
first several detection (\code{p}) estimates.
}
\author{
Ryan Kinzer
}
